Todo-list
Started the project's script. Now the user is able to add a task, delete it and mark as done. Plans to make the task editable, to unmark as done, to show just the done tasks or just the todo tasks. Also, I'll try to implement a small database feature to save the tasks and its status, rather than a python dictionary or a .json file